# Improve performance and parallelism, and reduce allocations in Planner