Improves performance and parallelism, and reduces allocations in **Planner** by:
* Eliminating virtual method calls.
* Using `List<T>.ForEach(Action<T> action)` instead of `ExtensionsForIEnumerableOfT.Map<T>(this IEnumerable<T> series, Action<T> action)`.
* Executing strategies outside of write lock.
* Removing duplicate null check in `Planner.CreateEmptyPlan(Type type)`.

** Before:**

|                            Method |        Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------------- |------------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|           GetPlan_Existing_Serial |    40.68 ns |  0.0496 ns |  0.0464 ns |           - |           - |           - |                   - |
|         GetPlan_Existing_Parallel |   263.68 ns |  5.2703 ns | 14.7786 ns |           - |           - |           - |                   - |
|      GetPlan_New_WithDelay_Serial |   407.55 ns |  1.2374 ns |  1.0969 ns |      0.0608 |      0.0304 |      0.0136 |               257 B |
|   GetPlan_New_WithoutDelay_Serial |   367.22 ns |  1.7601 ns |  1.5603 ns |      0.0608 |      0.0304 |      0.0136 |               257 B |
|    GetPlan_New_WithDelay_Parallel | 1,289.47 ns | 25.2864 ns | 34.6123 ns |      0.0543 |      0.0260 |      0.0087 |                 4 B |
| GetPlan_New_WithoutDelay_Parallel | 1,233.33 ns | 33.7178 ns | 33.1154 ns |      0.0543 |      0.0260 |      0.0087 |                 4 B |

**After:**

|                            Method |        Mean |      Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------------- |------------:|-----------:|----------:|------------:|------------:|------------:|--------------------:|
|           GetPlan_Existing_Serial |    40.46 ns |  0.8365 ns |  1.117 ns |           - |           - |           - |                   - |
|         GetPlan_Existing_Parallel |   193.65 ns |  3.8201 ns |  7.175 ns |           - |           - |           - |                   - |
|      GetPlan_New_WithDelay_Serial |   347.70 ns |  2.3334 ns |  2.183 ns |      0.0505 |      0.0250 |      0.0119 |               217 B |
|   GetPlan_New_WithoutDelay_Serial |   318.24 ns |  3.6420 ns |  3.229 ns |      0.0505 |      0.0250 |      0.0119 |               217 B |
|    GetPlan_New_WithDelay_Parallel | 1,102.92 ns | 21.9722 ns | 25.303 ns |      0.0477 |      0.0239 |      0.0109 |                 4 B |
| GetPlan_New_WithoutDelay_Parallel | 1,125.92 ns | 22.4928 ns | 49.372 ns |      0.0456 |      0.0217 |      0.0087 |                 4 B |